+++ docs/changelog.md
@@ -26,6 +26,7 @@ category: Administration
 * docs: add documentation for project specific estimate field configuration
 * feat: allow the card set to be overwritten per project (#101)
 * feat: allow to remove project specific configuration and go back to defaults
+* refactor: do not create empty project specific settings when not needed
 
 ### [20.10.0] - 2020-10-08